+++ packages/app/cypress/component/support/index.ts
@@ -26,8 +26,10 @@ import { createPinia } from '../../../src/store'
 import { setActivePinia } from 'pinia'
 import type { Pinia } from 'pinia'
 import 'cypress-real-events/support'
+import 'cypress-plugin-tab'
 
 import { installCustomPercyCommand } from '@packages/frontend-shared/cypress/support/customPercyCommand'
+import { tabUntil } from '@packages/frontend-shared/cypress/support/tab-until'
 
 let pinia: Pinia
 
@@ -50,3 +52,4 @@ registerMountFn({ plugins: [() => createRouter(), () => pinia] })
 installCustomPercyCommand()
 
 Cypress.on('uncaught:exception', (err) => !err.message.includes('ResizeObserver loop limit exceeded'))
+Cypress.Commands.add('tabUntil', tabUntil)