+++ packages/app/src/specs/InlineSpecList.vue
@@ -9,7 +9,9 @@
     <InlineSpecListHeader
       v-model:specFilterModel="specFilterModel"
       :result-count="specs.length"
+      :is-run-all-specs-allowed="runAllSpecsStore.isRunAllSpecsAllowed"
       @newSpec="showModal = true"
+      @run-all-specs="runAllSpecsStore.runAllSpecs"
     />
     <InlineSpecListTree
       :specs="specs"
@@ -29,8 +31,9 @@ import InlineSpecListHeader from './InlineSpecListHeader.vue'
 import InlineSpecListTree from './InlineSpecListTree.vue'
 import CreateSpecModal from './CreateSpecModal.vue'
 import { fuzzySortSpecs, makeFuzzyFoundSpec, useCachedSpecs } from './spec-utils'
-import type { FuzzyFoundSpec } from './spec-utils'
+import type { FuzzyFoundSpec } from './tree/useCollapsibleTree'
 import { useSpecFilter } from '../composables/useSpecFilter'
+import { useRunAllSpecsStore } from '../store/run-all-specs-store'
 
 gql`
 fragment SpecNode_InlineSpecList on Spec {
@@ -79,5 +82,7 @@ const specs = computed<FuzzyFoundSpec[]>(() => {
 
   return fuzzySortSpecs(specs, debouncedSpecFilterModel.value)
 })
+
+const runAllSpecsStore = useRunAllSpecsStore()
 
 </script>