+++ packages/app/src/specs/InlineSpecListTree.vue
@@ -18,12 +18,12 @@
         relative"
         data-cy="spec-row-item"
         :data-selected-spec="isCurrentSpec(row.data)"
-        @click.self="submit(row.data, row.index)"
+        @click.self="submitOrToggle(row.data, row.index)"
       >
-        <RouterLink
+        <component
+          :is="row.data.isLeaf ? RouterLink : 'button'"
           :ref="el => setItemRef(el, row.index)"
           :key="row.data.data?.absolute"
-          :tabindex="isTabbable(row, row.index) ? '0' : '-1'"
           :style="{ paddingLeft: `${(row.data.depth - 2) * 10 + 16}px` }"
           class="border-transparent outline-none border-1 w-full group focus-visible:bg-gray-900 before:(border-r-4 border-transparent h-28px rounded-r-4px absolute left-[-4px] w-8px) "
           :class="{
@@ -31,9 +31,9 @@
             'before:focus:border-r-indigo-300 before:focus-visible:border-r-transparent before:hover:border-r-indigo-300': !isCurrentSpec(row.data)
           }"
           :to="{ path: '/specs/runner', query: { file: row.data.data?.relative?.replace(/\\/g, '/') } }"
+          :aria-expanded="row.data.isLeaf ? null : row.data.expanded"
           @focus="resetFocusIfNecessary(row, row.index)"
-          @click.capture.prevent="submit(row.data, row.index)"
-          @keydown.enter.space.prevent.stop="submit(row.data, row.index)"
+          @click.prevent="submitOrToggle(row.data, row.index)"
           @keydown.left.right.prevent.stop="toggle(row.data, row.index)"
         >
           <SpecFileItem
@@ -52,26 +52,35 @@
             :expanded="treeSpecList[row.index].expanded.value"
             :indexes="row.data.highlightIndexes"
             data-cy="directory-item"
-          />
-        </RouterLink>
+          >
+            <template #run-all-specs>
+              <InlineRunAllSpecs
+                v-if="runAllSpecsStore.isRunAllSpecsAllowed"
+                :directory="row.data.name"
+                class="flex h-full opacity-0 run-all justify-center items-center"
+                :spec-number="runAllSpecsStore.directoryChildren[row.data.id].length"
+                @runAllSpecs="() => runAllSpecsStore.runSelectedSpecs(row.data.id)"
+              />
+            </template>
+          </DirectoryItem>
+        </component>
       </li>
     </ul>
   </div>
 </template>
 
 <script setup lang="ts">
-import { useCollapsibleTree } from '@packages/frontend-shared/src/composables/useCollapsibleTree'
-import type { UseCollapsibleTreeNode } from '@packages/frontend-shared/src/composables/useCollapsibleTree'
-import { buildSpecTree } from './spec-utils'
-import type { SpecTreeNode, FuzzyFoundSpec } from './spec-utils'
+import { UseCollapsibleTreeNode, useCollapsibleTree, SpecTreeNode, FuzzyFoundSpec, buildSpecTree } from './tree/useCollapsibleTree'
 import SpecFileItem from './SpecFileItem.vue'
 import { computed, watch, onMounted } from 'vue'
 import DirectoryItem from './DirectoryItem.vue'
 import { RouterLink, useRouter } from 'vue-router'
 import { useSpecStore } from '../store'
-import { useVirtualList } from '@packages/frontend-shared/src/composables/useVirtualList'
-import { useVirtualListNavigation } from '@packages/frontend-shared/src/composables/useVirtualListNavigation'
+import { useVirtualList } from './tree/useVirtualList'
+import { useVirtualListNavigation } from './tree/useVirtualListNavigation'
 import { useStudioStore } from '../store/studio-store'
+import InlineRunAllSpecs from './InlineRunAllSpecs.vue'
+import { useRunAllSpecsStore } from '../store/run-all-specs-store'
 
 const props = defineProps<{
   specs: FuzzyFoundSpec[]
@@ -115,7 +124,7 @@ const toggle = (row: UseCollapsibleTreeNode<SpecTreeNode<FuzzyFoundSpec>>, idx: 
   row.toggle()
 }
 
-const submit = (row: UseCollapsibleTreeNode<SpecTreeNode<FuzzyFoundSpec>>, idx: number) => {
+const submitOrToggle = (row: UseCollapsibleTreeNode<SpecTreeNode<FuzzyFoundSpec>>, idx: number) => {
   // If the user selects a new spec while in studio mode, turn studio mode off
   const studioStore = useStudioStore()
 
@@ -156,6 +165,12 @@ const resetFocusIfNecessary = (row, index) => {
   }
 }
 
+const runAllSpecsStore = useRunAllSpecsStore()
+
+watch(collapsible, () => {
+  runAllSpecsStore.setRunAllSpecsData(collapsible.value.tree)
+}, { immediate: true })
+
 </script>
 
 <style scoped>
@@ -167,6 +182,11 @@ a::before {
 /** Header is 64px, padding-top is 8px **/
 .specs-list-container {
   height: calc(100vh - 64px - 8px);
+}
+
+/** For run all specs group hover to work */
+[data-cy=spec-row-item]:hover .run-all, [data-cy=spec-row-item]:focus-within .run-all {
+  opacity: 1 !important;
 }
 
 </style>