+++ packages/server/lib/open_project.ts
@@ -94,7 +94,6 @@ export class OpenProject {
       isTextTerminal: !!cfg.isTextTerminal,
       downloadsFolder: cfg.downloadsFolder,
       experimentalModifyObstructiveThirdPartyCode: cfg.experimentalModifyObstructiveThirdPartyCode,
-      experimentalSessionAndOrigin: cfg.experimentalSessionAndOrigin,
       experimentalWebKitSupport: cfg.experimentalWebKitSupport,
       ...prevOptions || {},
     }